update to have styling
