+++ app/src/app/login-page/page.tsx
@@ -1,5 +1,7 @@
 'use client'
 
+import "../../globals.scss";
+import "../../styles/main.scss";
 import React, { useCallback, useState } from "react";
 import Layout from "../../components/common/layout/Layout";
 import { useNavigate } from "react-router-dom";
