+++ app/src/services/charts/BarchartService.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { BarchartProps } from "../../data/interfaces/BarchartData";
 import {
   initializeSvg,
